points pool should be textgrid-level
